# Mon_Miam_Miam/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon Miam Miam API",
        "description": "API for Mon Miam Miam application",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/complaints": {
            "get": {
                "tags": [
                    "Complaints"
                ],
                "summary": "Lister les réclamations (pour le personnel)",
                "description": "Retourne une liste paginée des réclamations. Accessible par les employés, managers et admins.",
                "operationId": "c896b092927624ef9ad25851388b1a78",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrer par statut (pending, in_progress, resolved, rejected)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des réclamations"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Complaints"
                ],
                "summary": "Créer une nouvelle réclamation (étudiant)",
                "operationId": "ccdda53fcb7c24fba177380d20aa10d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id",
                                    "description"
                                ],
                                "properties": {
                                    "order_id": {
                                        "description": "ID de la commande concernée",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "description": "Description détaillée de la réclamation",
                                        "type": "string",
                                        "example": "Le plat était froid."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Réclamation créée avec succès"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/complaints/my": {
            "get": {
                "tags": [
                    "Complaints"
                ],
                "summary": "Lister mes réclamations (étudiant)",
                "description": "Retourne la liste paginée des réclamations de l'utilisateur authentifié.",
                "operationId": "be8e1c210d10eb9c4218e8cdc2172259",
                "responses": {
                    "200": {
                        "description": "Liste de mes réclamations"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/complaints/{complaint}": {
            "get": {
                "tags": [
                    "Complaints"
                ],
                "summary": "Afficher les détails d'une réclamation",
                "operationId": "5ba6f8b1fd1233f80f645efc71a5cfc5",
                "parameters": [
                    {
                        "name": "complaint",
                        "in": "path",
                        "description": "ID de la réclamation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la réclamation"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Réclamation non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/complaints/{complaint}/respond": {
            "post": {
                "tags": [
                    "Complaints"
                ],
                "summary": "Répondre à une réclamation (employé/manager)",
                "operationId": "2e3ec68a0abcd9c89adcbd17e6818183",
                "parameters": [
                    {
                        "name": "complaint",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "response"
                                ],
                                "properties": {
                                    "response": {
                                        "description": "Réponse de l'employé",
                                        "type": "string",
                                        "example": "Nous sommes désolés pour ce désagrément."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réponse enregistrée"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/complaints/{complaint}/resolve": {
            "patch": {
                "tags": [
                    "Complaints"
                ],
                "summary": "Résoudre ou rejeter une réclamation (admin/manager)",
                "operationId": "8d51b4e3baf1b39e4e5456117b22bdd0",
                "parameters": [
                    {
                        "name": "complaint",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "Nouveau statut",
                                        "type": "string",
                                        "enum": [
                                            "resolved",
                                            "rejected"
                                        ]
                                    },
                                    "response": {
                                        "description": "Réponse finale (optionnelle si déjà présente)",
                                        "type": "string",
                                        "example": "La réclamation a été traitée."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réclamation mise à jour"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}